+++ app/src/main/java/com/app/arcabyolimpo/data/remote/dto/password/RecoverPasswordDto.kt
@@ -1,0 +1,7 @@
+package com.app.arcabyolimpo.data.remote.dto.password
+
+import com.google.gson.annotations.SerializedName
+
+data class RecoverPasswordDto (
+    @SerializedName("email") val email: String
+) {}