+++ app/src/main/java/com/app/arcabyolimpo/di/AppModule.kt
@@ -7,9 +7,15 @@ import com.app.arcabyolimpo.data.remote.interceptor.AuthInterceptor
 import com.app.arcabyolimpo.data.remote.interceptor.SessionManager
 import com.app.arcabyolimpo.data.remote.interceptor.TokenAuthenticator
 import com.app.arcabyolimpo.data.repository.auth.UserRepositoryImpl
+import com.app.arcabyolimpo.data.repository.password.PasswordPasswordUserRepositoryImpl
+import com.app.arcabyolimpo.data.repository.supplies.SupplyRepositoryImpl
+import com.app.arcabyolimpo.data.repository.workshops.WorkshopRepositoryImpl
 import com.app.arcabyolimpo.domain.repository.auth.UserRepository
 import com.app.arcabyolimpo.data.repository.ExternalCollaborator.ExternalCollabRepositoryImpl
 import com.app.arcabyolimpo.domain.repository.ExternalCollabRepository.ExternalCollabRepository
+import com.app.arcabyolimpo.domain.repository.password.PasswordUserRepository
+import com.app.arcabyolimpo.domain.repository.supplies.SupplyRepository
+import com.app.arcabyolimpo.domain.repository.workshops.WorkshopRepository
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -93,4 +99,44 @@ object AppModule {
     fun provideExternalCollabRepository(
         api: ArcaApi,
     ): ExternalCollabRepository = ExternalCollabRepositoryImpl(api)
+
+    @Provides
+    @Singleton
+    fun providePasswordUserRepository(
+        api: ArcaApi
+    ): PasswordUserRepository = PasswordPasswordUserRepositoryImpl(api)
+
+    /**
+     * Provides the [SupplyRepository] implementation.
+     *
+     * This repository handles all supply-related data operations,
+     * including fetching the supply list and retrieving detailed
+     * information for a specific supply. It uses [ArcaApi] as the
+     * remote data source and maps API responses to domain models.
+     *
+     * @param api The [ArcaApi] instance used to perform network requests.
+     * @return A singleton instance of [SupplyRepositoryImpl].
+     */
+    @Provides
+    @Singleton
+    fun provideSupplyRepository(
+        api: ArcaApi
+    ): SupplyRepository = SupplyRepositoryImpl(api)
+
+    /**
+     * Provides the [WorkshopRepository] implementation.
+     *
+     * This repository handles all workshop-related data operations,
+     * including fetching the workshop list and retrieving detailed
+     * information for a specific workshop. It uses [ArcaApi] as the
+     * remote data source and maps API responses to domain models.
+     *
+     * @param api The [ArcaApi] instance used to perform network requests.
+     * @return A singleton instance of [WorkshopRepositoryImpl].
+     */
+    @Provides
+    @Singleton
+    fun provideWorkshopRepository(
+        api: ArcaApi
+    ): WorkshopRepository = WorkshopRepositoryImpl(api)
 }