+++ app/src/main/java/com/app/arcabyolimpo/presentation/navigation/NavGraph.kt
@@ -1,23 +1,39 @@
 package com.app.arcabyolimpo.presentation.navigation
 
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
+import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
 import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
+import androidx.navigation.navArgument
+import androidx.navigation.navDeepLink
 import com.app.arcabyolimpo.data.remote.interceptor.SessionManager
 import com.app.arcabyolimpo.domain.model.auth.UserRole
-import com.app.arcabyolimpo.presentation.screens.admin.CoordinatorHomeScreen
-import com.app.arcabyolimpo.presentation.screens.client.CollaboratorHomeScreen
+import com.app.arcabyolimpo.presentation.common.components.LoadingShimmer
+import com.app.arcabyolimpo.presentation.screens.accountactivation.AccountActivationScreen
+import com.app.arcabyolimpo.presentation.screens.home.assistant.CollaboratorHomeScreen
+import com.app.arcabyolimpo.presentation.screens.home.coordinator.CoordinatorHomeScreen
 import com.app.arcabyolimpo.presentation.screens.login.LoginScreen
+import com.app.arcabyolimpo.presentation.screens.passwordrecovery.PasswordRecoveryScreen
+import com.app.arcabyolimpo.presentation.screens.passwordregisteration.PasswordRegistrationScreen
+import com.app.arcabyolimpo.presentation.screens.passwordregisteration.PasswordRegistrationSuccessScreen
 import com.app.arcabyolimpo.presentation.screens.splash.SplashScreen
 import com.app.arcabyolimpo.presentation.screens.ExternalCollab.ExternalCollabList.ExternalCollabListScreen
 import com.app.arcabyolimpo.presentation.screens.ExternalCollab.ExternalCollabDetail.ExternalCollabDetailScreen
 import com.app.arcabyolimpo.presentation.screens.ExternalCollab.RegisterExternalCollab.ExternalCollabRegisterScreen
+import com.app.arcabyolimpo.presentation.screens.supply.SupplyListScreen
+import com.app.arcabyolimpo.presentation.screens.tokenverification.TokenVerificationFailedScreen
+import com.app.arcabyolimpo.presentation.screens.tokenverification.TokenVerificationViewModel
+import com.app.arcabyolimpo.presentation.screens.workshop.AddNewWorkshopScreen
+import com.app.arcabyolimpo.presentation.screens.workshop.WorkshopsListScreen
 
 /**
  * Defines all available destinations (routes) in the application.
@@ -31,9 +47,9 @@ sealed class Screen(
 
     object Login : Screen("login")
 
-    object CoordinatorHome : Screen("admin")
+    object PasswordRecovery : Screen("password-recovery")
 
-    object CollaboratorHome : Screen("client")
+    object AccountActivation : Screen("account-activation")
 
     object ExternalCollabList : Screen("external_collab_list")
 
@@ -43,6 +59,27 @@ sealed class Screen(
 
     object ExternalCollabRegister : Screen("external_collab_register")
 
+    object TokenVerification : Screen("user/verify-token?token={token}") {
+        fun createRoute(token: String) = "user/verify-token?token=$token"
+    }
+
+    object TokenVerificationFailed : Screen("token-activation-failed")
+
+    object PasswordRegistration : Screen("password-registration/{email}") {
+        fun createRoute(email: String) = "password-registration/$email"
+    }
+
+    object PasswordRegistrationSuccess : Screen("pasword-registration-success")
+
+    object CoordinatorHome : Screen("coordinator")
+
+    object CollaboratorHome : Screen("collaborator")
+
+    object SuppliesList : Screen("supply")
+
+    object WorkshopsList : Screen("workshop")
+
+    object AddNewWorkshop: Screen("workshop/add")
 }
 
 /**
@@ -74,21 +111,26 @@ fun ArcaNavGraph(
         }
     }
 
+
     /** Defines all navigation. The start destination is the Splash screen. */
     NavHost(
         navController = navController,
-        startDestination = Screen.ExternalCollabList.route,
+        startDestination = Screen.Splash.route,
         modifier = modifier,
     ) {
         /** Splash Screen */
         composable(Screen.Splash.route) {
             SplashScreen(onNavigate = { role ->
                 when (role) {
-                    UserRole.COORD ->
+                    UserRole.COORDINADOR ->
                         navController.navigate(Screen.CoordinatorHome.route) {
                             popUpTo(Screen.Splash.route) { inclusive = true }
                         }
-                    UserRole.COLAB ->
+                    UserRole.ASISTENTE ->
+                        navController.navigate(Screen.CollaboratorHome.route) {
+                            popUpTo(Screen.Splash.route) { inclusive = true }
+                        }
+                    UserRole.BECARIO ->
                         navController.navigate(Screen.CollaboratorHome.route) {
                             popUpTo(Screen.Splash.route) { inclusive = true }
                         }
@@ -102,18 +144,146 @@ fun ArcaNavGraph(
 
         /** Login Screen */
         composable(Screen.Login.route) {
-            LoginScreen(onLoginSuccess = { role ->
-                when (role) {
-                    UserRole.COORD ->
-                        navController.navigate(Screen.CoordinatorHome.route) {
-                            popUpTo(Screen.Login.route) { inclusive = true }
-                        }
-                    UserRole.COLAB ->
-                        navController.navigate(Screen.CollaboratorHome.route) {
-                            popUpTo(Screen.Login.route) { inclusive = true }
-                        }
+            LoginScreen(
+                onLoginSuccess = { role ->
+                    when (role) {
+                        UserRole.COORDINADOR ->
+                            navController.navigate(Screen.CoordinatorHome.route) {
+                                popUpTo(Screen.Login.route) { inclusive = true }
+                            }
+                        UserRole.ASISTENTE ->
+                            navController.navigate(Screen.CollaboratorHome.route) {
+                                popUpTo(Screen.Login.route) { inclusive = true }
+                            }
+                        UserRole.BECARIO ->
+                            navController.navigate(Screen.CollaboratorHome.route) {
+                                popUpTo(Screen.Login.route) { inclusive = true }
+                            }
+                    }
+                },
+                onRecoverPasswordClick = {
+                    navController.navigate(Screen.PasswordRecovery.route)
+                },
+                onAccountActivationClick = {
+                    navController.navigate(Screen.AccountActivation.route)
+                },
+            )
+        }
+
+        composable(Screen.PasswordRecovery.route) {
+            PasswordRecoveryScreen(
+                onBackClick = {
+                    if (!navController.popBackStack()) {
+                        navController.navigate(Screen.Login.route)
+                    }
+                },
+                viewModel = hiltViewModel(),
+            )
+        }
+
+        composable(Screen.AccountActivation.route) {
+            AccountActivationScreen(
+                onBackClick = {
+                    if (!navController.popBackStack()) {
+                        navController.navigate(Screen.Login.route)
+                    }
+                },
+                viewModel = hiltViewModel(),
+            )
+        }
+
+        composable(
+            route = Screen.TokenVerification.route,
+            arguments =
+                listOf(
+                    navArgument("token") {
+                        type = NavType.StringType
+                        nullable = true
+                    },
+                ),
+            deepLinks =
+                listOf(
+                    navDeepLink {
+                        uriPattern = "arcabyolimpo://user/verify-token?token={token}"
+                    },
+                ),
+        ) { backStackEntry ->
+            val token = backStackEntry.arguments?.getString("token")
+            val viewModel: TokenVerificationViewModel = hiltViewModel()
+            val uiState by viewModel.uiState.collectAsState()
+
+            LaunchedEffect(token) {
+                token?.let { viewModel.getTokenVerification(it) }
+            }
+
+            when {
+                uiState.response?.valid == true -> {
+                    PasswordRegistrationScreen(
+                        email = uiState.response?.email,
+                        onBackClick = { navController.popBackStack() },
+                        onPasswordRegistrationSucessClick = {
+                            navController.navigate(Screen.PasswordRegistrationSuccess.route)
+                        },
+                        viewModel = hiltViewModel(),
+                    )
                 }
-            })
+
+                uiState.isLoading -> {
+                    LoadingShimmer(
+                        modifier = Modifier
+                            .fillMaxSize()
+                    )
+                }
+
+                uiState.error == "Invalid or expired token" -> {
+                    TokenVerificationFailedScreen(onBackClick = { navController.popBackStack() })
+                }
+
+                else -> {
+                    LoadingShimmer(
+                        modifier = Modifier
+                            .fillMaxSize()
+                    )
+                }
+            }
+        }
+
+        composable(Screen.TokenVerificationFailed.route) {
+            TokenVerificationFailedScreen(
+                onBackClick = {
+                    if (!navController.popBackStack()) {
+                        navController.navigate(Screen.Login.route)
+                    }
+                },
+            )
+        }
+
+        composable(
+            route = Screen.PasswordRegistration.route,
+            arguments = listOf(navArgument("email") { type = NavType.StringType }),
+        ) { backStackEntry ->
+            val email = backStackEntry.arguments?.getString("email")
+            PasswordRegistrationScreen(
+                email = email,
+                onBackClick = {
+                    if (!navController.popBackStack()) {
+                        navController.navigate(Screen.Login.route)
+                    }
+                },
+                onPasswordRegistrationSucessClick = {
+                    navController.popBackStack()
+                    navController.popBackStack()
+                },
+                viewModel = hiltViewModel(),
+            )
+        }
+
+        composable(Screen.PasswordRegistrationSuccess.route) {
+            PasswordRegistrationSuccessScreen(
+                onBackClick = {
+                    navController.navigate(Screen.Login.route)
+                },
+            )
         }
 
         /** Coordinator Home Screen */
@@ -161,6 +331,64 @@ fun ArcaNavGraph(
                 onSuccess = {
                     navController.popBackStack()
                 }
+            )
+        }
+
+        /**
+         * Workshops List Screen.
+         *
+         * This composable represents the screen where users can view and interact with
+         * the list of available workshops.
+         *
+         * It connects to the [WorkshopsListScreen] composable, which displays the UI and
+         * interacts with its corresponding [WorkshopsListViewModel] to handle data fetching,
+         * loading states, and errors.
+         *
+         */
+        composable(Screen.WorkshopsList.route) {
+            WorkshopsListScreen(
+                navController = navController,
+                workshopClick = {}
+            )
+        }
+
+        /**
+         * Workshops Add Screen.
+         *
+         * This composable represents the screen where users can view and interact with
+         * the register of a new workshop.
+         *
+         * It connects to the [AddNewWorkshopScreen] composable, which displays the UI and
+         * interacts with its corresponding [AddNewWorkshopViewModel] to handle data fetching,
+         * loading states, and errors.
+         *
+         */
+        composable(Screen.AddNewWorkshop.route) {
+            AddNewWorkshopScreen(
+                navController = navController,
+                viewModel = hiltViewModel(),
+                onSuccess = {
+                    navController.popBackStack()
+                }
+            )
+        }
+
+        /**
+         * Supply List Screen.
+         *
+         * This composable represents the screen where users can view and interact with
+         * the list of available supplies.
+         *
+         * It connects to the [SupplyListScreen] composable, which displays the UI and
+         * interacts with its corresponding [SuppliesListViewModel] to handle data fetching,
+         * loading states, and errors.
+         *
+         */
+        composable(Screen.SuppliesList.route) {
+            SupplyListScreen(
+                onSupplyClick = { id ->
+                    navController.navigate("supplyDetail/$id")
+                },
             )
         }
     }