+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/accountactivation/AccountActivationViewModel.kt
@@ -1,0 +1,51 @@
+package com.app.arcabyolimpo.presentation.screens.accountactivation
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.app.arcabyolimpo.domain.common.Result
+import com.app.arcabyolimpo.domain.usecase.password.PostPasswordRecoveryUseCase
+import com.app.arcabyolimpo.presentation.screens.passwordrecovery.PasswordRecoveryUiState
+import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.update
+import kotlinx.coroutines.launch
+import javax.inject.Inject
+
+@HiltViewModel
+class AccountActivationViewModel
+@Inject
+constructor(
+    private val postPasswordRecoveryUseCase: PostPasswordRecoveryUseCase,
+) : ViewModel() {
+    private val _uiState = MutableStateFlow(PasswordRecoveryUiState())
+    val uiState: StateFlow<PasswordRecoveryUiState> = _uiState.asStateFlow()
+
+    fun postPasswordRecovery(email: String) {
+        viewModelScope.launch {
+            postPasswordRecoveryUseCase(email).collect { result ->
+                _uiState.update { state ->
+                    when (result) {
+                        is Result.Loading ->
+                            state.copy(
+                                isLoading = true,
+                            )
+                        is Result.Success ->
+                            state.copy(
+                                message = result.data,
+                                isLoading = false,
+                                error = null,
+                            )
+                        is Result.Error ->
+                            state.copy(
+                                error = result.exception.message,
+                                isLoading = false,
+                            )
+                    }
+                }
+
+            }
+        }
+    }
+}