+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/home/assistant/CollaboratorHomeScreen.kt
@@ -1,4 +1,4 @@
-package com.app.arcabyolimpo.presentation.screens.admin
+package com.app.arcabyolimpo.presentation.screens.home.assistant
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -18,7 +18,7 @@ import com.app.arcabyolimpo.presentation.screens.session.SessionViewModel
 
 @Suppress("ktlint:standard:function-naming")
 @Composable
-fun CoordinatorHomeScreen() {
+fun CollaboratorHomeScreen() {
     val sessionViewModel: SessionViewModel = hiltViewModel()
 
     Column(
@@ -29,7 +29,7 @@ fun CoordinatorHomeScreen() {
         horizontalAlignment = Alignment.CenterHorizontally,
         verticalArrangement = Arrangement.Center,
     ) {
-        Text("Bienvenido Coordinador 👑", style = MaterialTheme.typography.headlineSmall)
+        Text("Bienvenido Colaborador 👑", style = MaterialTheme.typography.headlineSmall)
         Spacer(modifier = Modifier.height(20.dp))
         Button(onClick = { sessionViewModel.logout() }) {
             Text("Cerrar sesión")