+++ app/src/main/java/com/app/arcabyolimpo/presentation/ui/components/atoms/buttons/AddButton.kt
@@ -1,32 +1,34 @@
 package com.app.arcabyolimpo.presentation.ui.components.atoms.buttons
 
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Icon
-import androidx.compose.material3.LargeFloatingActionButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.app.arcabyolimpo.presentation.theme.Poppins
+import com.app.arcabyolimpo.presentation.ui.components.atoms.icons.PlusIcon
 import com.app.arcabyolimpo.ui.theme.PrimaryBlue
 import com.app.arcabyolimpo.ui.theme.White
 
 /**
+ * A circular "Add" button component that displays a "+" symbol.
  *
+ * @param modifier Modifier to adjust the button's layout (size, padding, etc.).
+ * @param onClick Lambda executed when the button is pressed.
+ *
+ * The button:
+ * - Uses a circular shape.
+ * - Has a white background.
+ * - Displays a large "+" in the primary blue color.
  */
+@Suppress("ktlint:standard:function-naming")
 @Composable
 fun AddButton(
     modifier: Modifier = Modifier,
@@ -34,31 +36,25 @@ fun AddButton(
 ) {
     Button(
         onClick = onClick,
-        modifier = modifier.size(56.dp),
+        modifier = modifier.size(72.dp),
         shape = CircleShape,
-        colors =
-            ButtonDefaults.buttonColors(
-                containerColor = White,
-            ),
-        contentPadding = PaddingValues(0.dp),
+        contentPadding = ButtonDefaults.ContentPadding,
+        colors = ButtonDefaults.buttonColors(
+            containerColor = White,
+        ),
     ) {
-        Text(
-            text = "+",
-            color = PrimaryBlue,
-            fontFamily = Poppins,
-            fontWeight = FontWeight.SemiBold,
-            fontSize = 24.sp,
+        PlusIcon(
+            tint = PrimaryBlue,
+            modifier = Modifier.size(32.dp),
         )
     }
 }
 
+@Suppress("ktlint:standard:function-naming")
 @Preview(showBackground = true)
 @Composable
 fun AddButtonPreview() {
     MaterialTheme {
-        AddButton(
-            modifier = Modifier.size(width = 95.dp, height = 25.dp),
-            onClick = {},
-        )
+        AddButton(onClick = {})
     }
 }